Refactor repo with filename convention and folder convention
