+++ datastructures/hashmap/hashmap.go
