+++ datastructures/linkedlist/singlelinkedlist/single-linkedlist.go
