+++ datastructures/linkedlist/singlylinkedlist.go
@@ -1,4 +1,4 @@
-package main
+package linkedlist
 
 import "fmt"
