+++ datastructures/trie/trie.go
