+++ graphs/floydwarshall.go
@@ -1,7 +1,7 @@
 // Floyd-Warshall algorithm
 // https://en.wikipedia.org/wiki/Floyd%E2%80%93Warshall_algorithm
 
-package main
+package graphs
 
 import (
 	"math"
