+++ math/gcd/gcd.go
