+++ math/power/fastexponent.go
