+++ math/power/fastexponent_test.go
