+++ math/primecheck/primecheck.go
