+++ math/primecheck/primecheck_test.go
