+++ searches/linearsearch.go
@@ -1,4 +1,4 @@
-package main
+package searches
 
 import "fmt"
