+++ sorts/topologicalsort.go
@@ -1,7 +1,7 @@
 //Package sorts a package for demonstrating sorting algorithms in Go
 package sorts
 
-/*type Dependency struct {
+type Dependency struct {
 	nodeId   int
 	children []int
 }
@@ -81,7 +81,7 @@ func topologicalSort(input []Edge) ([]int, [][]int) {
 		levels = append(levels, level)
 	}
 	return res, levels
-}*/
+}
 
 /*func main() {
 	var edges []Edge
